Remove legacy cruft, leave only current (MVP) API
