--- backend/MVP_final_API/db.py
@@ -1,3 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy() # will link to our flask app to elements in a database
