--- backend/api/resources/__init__.py
@@ -1,1 +1,0 @@
-# this __init__file makes the resources folder a "package", meaning things can be imported from this folder.
