--- backend/api/security.py
@@ -1,10 +1,0 @@
-from models.user import UserModel
-
-def authenticate(username, password):
-    user = UserModel.find_by_username(username) # user == None if no match
-    if user and user.password == password:
-        return user
-
-def identity(payload):
-    user_id = payload["identity"]
-    return UserModel.find_by_id(user_id)
